Flaskの除去とNative Tornado化
動作確認済み(Android Firefox)
